mostlyComplete

+++ frontend/src/screens/SignupScreen.jsx
@@ -8,33 +8,34 @@ import { useContext, useState } from "react";
 import { Store } from "../Store";
 import { toast } from "react-toastify";
 
-export default function SignInScreen() {
+export default function SignupScreen() {
   const navigate = useNavigate();
   const { search } = useLocation();
   const redirectInUrl = new URLSearchParams(search).get("redirect");
   const redirect = redirectInUrl ? redirectInUrl : "/";
+  const [name , setName] = useState('');
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState('');
   const { state, dispatch: ctxDispatch } = useContext(Store);
-  const [showToast, setShowToast] = useState(false);
   const submitHandler = async (e) => {
     e.preventDefault();
+    if(password !== confirmPassword){
+        toast.error('Password do not match');
+        return ;
+    }
     try {
-      const { data } = await Axios.post("api/users/signin", {
+      const { data } = await Axios.post("api/users/signup", {
+        name,
         email,
         password,
       });
       ctxDispatch({ type: "USER_SIGNIN", payload: data });
       localStorage.setItem("userInfo", JSON.stringify(data));
-      navigate(redirect || "/");
+      navigate(redirect || '/');
       window.location.reload();
     } catch (error) {
-      const errorMessage =
-        error.response && error.response.data.message
-          ? error.response.data.message
-          : "Invalid Username or Password";
-          setShowToast(true); // Show the toast
-          toast.error(errorMessage);
+        toast.error(error.message);
     }
   };
 
@@ -42,10 +43,17 @@ export default function SignInScreen() {
     <>
       <Container className="small-container">
         <Helmet>
-          <title>Sign in</title>
+          <title>Sign Up</title>
         </Helmet>
-        <h1 className="my-3">Sign in</h1>
+        <h1 className="my-3">Sign Up</h1>
         <Form onSubmit={submitHandler}>
+        <Form.Group className="mb-3" controlId="name">
+            <Form.Label>Name</Form.Label>
+            <Form.Control
+              required
+              onChange={(e) => setName(e.target.value)}
+            ></Form.Control>
+          </Form.Group>
           <Form.Group className="mb-3" controlId="email">
             <Form.Label>Email</Form.Label>
             <Form.Control
@@ -62,13 +70,21 @@ export default function SignInScreen() {
               onChange={(e) => setPassword(e.target.value)}
             ></Form.Control>
           </Form.Group>
+          <Form.Group className="mb-3" controlId="confirmPassword">
+            <Form.Label>Confirm Password</Form.Label>
+            <Form.Control
+              type="password"
+              required
+              onChange={(e) => setConfirmPassword(e.target.value)}
+            ></Form.Control>
+          </Form.Group>
           <div className="mb-3">
-            <Button type="submit">Sign in</Button>
+            <Button type="submit">Sign Up</Button>
           </div>
           <div className="mb-3">
-            New Customer?{" "}
-            <RouterLink to={`/signup?redirect=${redirect}`}>
-              Create your account
+            Already have an account?{" "}
+            <RouterLink to={`/signin?redirect=${redirect}`}>
+              Sign-In
             </RouterLink>
           </div>
         </Form>
